requirements.txt
torch
torchvision
tensorboard
pandas
seaborn
scikit-learn
umap-learn
matplotlib
tqdm
